backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.1
mysqlclient==2.2.7
djangorestframework==3.15.2
pandas==2.2.3
tqdm==4.67.1
django-cors-headers==4.7.0
gunicorn==23.0.0
uvicorn==0.34.0
python-dotenv==1.0.1